Parser passes makeshift tests.
TODO: error checking, disambiguation, try/p

try/p is nice in that it helps disambiguate, but at the price of
losing quality in error reporting. Not only that, but try/p and
the existence of app/p really make it so that stuff overlaps.

This is not critical atm because parse.rkt and compile.rkt catch these
things, but we should implement arity checking in compile.rkt