Changed std.rot to std.lonn
In a way, .lonn are header files. Terrible, I know.
But due to the complete hackery that is mread, parsing it as its own
thing is not possible.

This is because `parse.rkt` receives the output of mread and assumes
that the last thing is an expression, for which we don't pattern match
function definitions. In other words, it misconstrues this as
an application of `define` and fails. It's not critical, and
in principle programs consisting of only definitions should be possible,
but what can you do amirite kids?